# host-server/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A plugin management system that supports dynamic loading and execution of plugins",
        "title": "Polyglot Plugin Host Server API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/plugins": {
            "get": {
                "description": "Get a list of all installed plugins with optional filters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Plugins"
                ],
                "summary": "List all plugins",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by namespace",
                        "name": "namespace",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by plugin type",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "active",
                            "inactive",
                            "disabled",
                            "error",
                            "installing"
                        ],
                        "type": "string",
                        "description": "Filter by plugin status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "linux",
                            "darwin",
                            "windows"
                        ],
                        "type": "string",
                        "description": "Filter by operating system",
                        "name": "os",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "amd64",
                            "arm64"
                        ],
                        "type": "string",
                        "description": "Filter by architecture",
                        "name": "arch",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Plugin"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
        "/api/plugins/install": {
            "post": {
                "description": "Install a plugin from a download URL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Plugins"
                ],
                "summary": "Install a new plugin",
                "parameters": [
                    {
                        "description": "Plugin installation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.InstallPluginRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Plugin"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
        "/api/plugins/{id}": {
            "get": {
                "description": "Get detailed information about a specific plugin by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Plugins"
                ],
                "summary": "Get plugin details",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Plugin ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Plugin"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove a plugin from the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Plugins"
                ],
                "summary": "Uninstall a plugin",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Plugin ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
        "/api/plugins/{id}/activate": {
            "post": {
                "description": "Activate a previously installed plugin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Plugins"
                ],
                "summary": "Activate a plugin",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Plugin ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
        "/api/plugins/{id}/call": {
            "post": {
                "description": "Execute a specific method on an active plugin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Plugins"
                ],
                "summary": "Call a plugin method",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Plugin ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Plugin call request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CallPluginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        },
        "/api/plugins/{id}/deactivate": {
            "post": {
                "description": "Deactivate an active plugin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Plugins"
                ],
                "summary": "Deactivate a plugin",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Plugin ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.AppError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "errors.AppError": {
            "type": "object",
            "properties": {
                "details": {
                    "description": "Additional details",
                    "type": "string"
                },
                "errorCode": {
                    "description": "Machine-readable error code",
                    "type": "string"
                },
                "message": {
                    "description": "Human-readable error message",
                    "type": "string"
                },
                "path": {
                    "description": "Request path",
                    "type": "string"
                },
                "success": {
                    "description": "Always false for errors",
                    "type": "boolean"
                },
                "timestamp": {
                    "description": "Unix timestamp",
                    "type": "integer"
                }
            }
        },
        "models.JSONMap": {
            "type": "object",
            "additionalProperties": {}
        },
        "models.Plugin": {
            "type": "object",
            "properties": {
                "arch": {
                    "description": "架构",
                    "type": "string"
                },
                "binary_path": {
                    "type": "string"
                },
                "config": {
                    "$ref": "#/definitions/models.JSONMap"
                },
                "created_at": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "download_url": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_used_at": {
                    "type": "integer"
                },
                "metadata": {
                    "description": "结构化元数据，存储 PluginMetadata",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.JSONMap"
                        }
                    ]
                },
                "name": {
                    "type": "string"
                },
                "namespace": {
                    "description": "命名空间",
                    "type": "string"
                },
                "os": {
                    "description": "操作系统",
                    "type": "string"
                },
                "protocol": {
                    "$ref": "#/definitions/models.PluginProtocol"
                },
                "protocol_version": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/models.PluginStatus"
                },
                "type": {
                    "$ref": "#/definitions/models.PluginType"
                },
                "updated_at": {
                    "type": "integer"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "models.PluginProtocol": {
            "type": "string",
            "enum": [
                "grpc",
                "net-rpc"
            ],
            "x-enum-comments": {
                "PluginProtocolGRPC": "gRPC 协议",
                "PluginProtocolNetRPC": "net/rpc 协议"
            },
            "x-enum-descriptions": [
                "gRPC 协议",
                "net/rpc 协议"
            ],
            "x-enum-varnames": [
                "PluginProtocolGRPC",
                "PluginProtocolNetRPC"
            ]
        },
        "models.PluginStatus": {
            "type": "string",
            "enum": [
                "active",
                "inactive",
                "disabled",
                "error",
                "installing"
            ],
            "x-enum-comments": {
                "PluginStatusActive": "激活",
                "PluginStatusDisabled": "已禁用",
                "PluginStatusError": "错误状态",
                "PluginStatusInactive": "未激活",
                "PluginStatusInstalling": "安装中"
            },
            "x-enum-descriptions": [
                "激活",
                "未激活",
                "已禁用",
                "错误状态",
                "安装中"
            ],
            "x-enum-varnames": [
                "PluginStatusActive",
                "PluginStatusInactive",
                "PluginStatusDisabled",
                "PluginStatusError",
                "PluginStatusInstalling"
            ]
        },
        "models.PluginType": {
            "type": "string",
            "enum": [
                "data-processing",
                "security",
                "integration",
                "extension"
            ],
            "x-enum-comments": {
                "PluginTypeDataProcessing": "数据处理：脱敏、转换、匿名化等",
                "PluginTypeExtension": "扩展：通用扩展功能",
                "PluginTypeIntegration": "集成：外部系统对接",
                "PluginTypeSecurity": "安全：加密、验证等"
            },
            "x-enum-descriptions": [
                "数据处理：脱敏、转换、匿名化等",
                "安全：加密、验证等",
                "集成：外部系统对接",
                "扩展：通用扩展功能"
            ],
            "x-enum-varnames": [
                "PluginTypeDataProcessing",
                "PluginTypeSecurity",
                "PluginTypeIntegration",
                "PluginTypeExtension"
            ]
        },
        "request.CallPluginRequest": {
            "type": "object",
            "required": [
                "id",
                "method",
                "params"
            ],
            "properties": {
                "id": {
                    "description": "Plugin ID from path parameter",
                    "type": "integer"
                },
                "method": {
                    "description": "Method name from request body",
                    "type": "string"
                },
                "params": {
                    "description": "Method parameters from request body",
                    "type": "object",
                    "additionalProperties": {}
                }
            }
        },
        "request.InstallPluginRequest": {
            "type": "object",
            "required": [
                "arch",
                "downloadURL",
                "name",
                "namespace",
                "os",
                "type",
                "version"
            ],
            "properties": {
                "arch": {
                    "description": "新增：架构",
                    "type": "string",
                    "enum": [
                        "amd64",
                        "arm64"
                    ]
                },
                "config": {
                    "$ref": "#/definitions/models.JSONMap"
                },
                "description": {
                    "type": "string"
                },
                "downloadURL": {
                    "type": "string"
                },
                "metadata": {
                    "$ref": "#/definitions/models.JSONMap"
                },
                "name": {
                    "type": "string"
                },
                "namespace": {
                    "description": "新增：命名空间",
                    "type": "string"
                },
                "os": {
                    "description": "新增：操作系统",
                    "type": "string",
                    "enum": [
                        "linux",
                        "darwin",
                        "windows"
                    ]
                },
                "type": {
                    "description": "修改：从枚举改为 string",
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        }
    },
    "tags": [
        {
            "description": "Plugin management operations",
            "name": "plugins"
        }
    ]
}